requirements.txt
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1671632566681/work
Automat @ file:///home/conda/feedstock_root/build_artifacts/automat_1667331175863/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bcrypt @ file:///home/conda/feedstock_root/build_artifacts/bcrypt_1666831323018/work
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1666764686339/work
certifi==2022.12.7
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1671179355965/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1678108872112/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1679481329611/work
constantly==15.1.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1681508576793/work
cssselect @ file:///home/conda/feedstock_root/build_artifacts/cssselect_1666980406338/work
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1680755454898/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dnspython==2.3.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
fake-useragent==1.1.3
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1680834529360/work
hyperlink @ file:///home/conda/feedstock_root/build_artifacts/hyperlink_1610092164190/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1681581693083/work
incremental @ file:///home/conda/feedstock_root/build_artifacts/incremental_1665859450441/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1679336319192/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
itemadapter @ file:///home/conda/feedstock_root/build_artifacts/itemadapter_1680243523209/work
itemloaders @ file:///home/conda/feedstock_root/build_artifacts/itemloaders_1661796597649/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
jmespath @ file:///home/conda/feedstock_root/build_artifacts/jmespath_1655568249366/work
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1681432441054/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1678994187655/work
lxml @ file:///home/conda/feedstock_root/build_artifacts/lxml_1671013475329/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
numpy==1.24.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.0.0
parsel @ file:///home/conda/feedstock_root/build_artifacts/parsel_1667392207259/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1679871349196/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
Protego @ file:///home/conda/feedstock_root/build_artifacts/protego_1647348122707/work
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1681775010077/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyDispatcher==2.0.5
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681142969746/work
pymongo==4.3.3
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1680037383858/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
pytz==2023.3
pyzmq @ file:///home/conda/feedstock_root/build_artifacts/pyzmq_1679316875704/work
queuelib @ file:///home/conda/feedstock_root/build_artifacts/queuelib_1629994290383/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1680286922386/work
requests-file @ file:///home/conda/feedstock_root/build_artifacts/requests-file_1591227832914/work
Scrapy @ file:///home/conda/feedstock_root/build_artifacts/scrapy_1678974838079/work
service-identity==18.1.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tldextract @ file:///home/conda/feedstock_root/build_artifacts/tldextract_1664943488272/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1666788591069/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
Twisted @ file:///home/conda/feedstock_root/build_artifacts/twisted_1667305486487/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
w3lib @ file:///Users/ktietz/demo/mc3/conda-bld/w3lib_1629359764703/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
zope.interface @ file:///home/conda/feedstock_root/build_artifacts/zope.interface_1679046859665/work
